v.13.0.0 - Refactoring
